make first example run
