+++ pennylane/tape/__init__.py
@@ -33,6 +33,7 @@ from .tapes import (
     QubitParamShiftTape,
     CVParamShiftTape,
     ReversibleTape,
+    BranchTape
 )
